requirements.txt
django>=2.1.2
mysqlclient==1.4.2
djangorestframework==3.9.2